#129 Final cleanup of Spark Streaming with Kafka and Cassandra demo.
