+++ spark-cassandra-connector-demos/src/main/scala/com/datastax/spark/connector/demo/streaming/embedded/Embedded.scala
@@ -1,12 +1,13 @@
-package com.datastax.spark.connector.streaming.kafka.embedded
+package com.datastax.spark.connector.demo.streaming.embedded
 
 import java.io.{IOException, File}
 import java.util.UUID
 
+import com.datastax.spark.connector.demo.Assertions
 import com.google.common.io.Files
 import org.apache.commons.lang3.SystemUtils
 import org.apache.spark.Logging
-import com.datastax.spark.connector.util.{Assertions, IOUtils}
+import com.datastax.spark.connector.util.IOUtils
 
 /** INTERNAL API.
   * Helper class for demos and tests. Useful for quick prototyping. */
